# backend/openapi2.json
{"openapi":"3.1.0","info":{"title":"NextGen Resume Intelligence System","description":"AI-powered resume analysis platform","version":"1.0.0"},"paths":{"/api/auth/register":{"post":{"tags":["Authentication"],"summary":"Register","description":"Register a new HR or Student account.","operationId":"register_api_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegister"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","description":"Login and receive a JWT access token.\nUse this token in the Authorization header for protected routes.","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/me":{"get":{"tags":["Authentication"],"summary":"Get Me","description":"Get the currently logged-in user's profile.\nRequires Authorization: Bearer <token>","operationId":"get_me_api_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/auth/logout":{"post":{"tags":["Authentication"],"summary":"Logout","description":"Logout — on the frontend, just delete the token from localStorage.\nJWT tokens are stateless so server-side logout just confirms the action.","operationId":"logout_api_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/hr/upload-resumes":{"post":{"tags":["HR"],"summary":"Upload Resumes","description":"Upload 1-100 resumes at once","operationId":"upload_resumes_api_hr_upload_resumes_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_resumes_api_hr_upload_resumes_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BulkUploadResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/hr/resumes":{"get":{"tags":["HR"],"summary":"Get Resumes","description":"Get all resumes uploaded by this HR user","operationId":"get_resumes_api_hr_resumes_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ResumeResponse"},"type":"array","title":"Response Get Resumes Api Hr Resumes Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/health":{"get":{"tags":["Health"],"summary":"Health Check","operationId":"health_check_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"tags":["Root"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_upload_resumes_api_hr_upload_resumes_post":{"properties":{"files":{"items":{"type":"string","contentMediaType":"application/octet-stream"},"type":"array","title":"Files"}},"type":"object","required":["files"],"title":"Body_upload_resumes_api_hr_upload_resumes_post"},"BulkUploadResponse":{"properties":{"total_uploaded":{"type":"integer","title":"Total Uploaded"},"successful":{"type":"integer","title":"Successful"},"duplicates":{"type":"integer","title":"Duplicates"},"failed":{"type":"integer","title":"Failed"},"resumes":{"items":{"$ref":"#/components/schemas/ResumeResponse"},"type":"array","title":"Resumes"}},"type":"object","required":["total_uploaded","successful","duplicates","failed","resumes"],"title":"BulkUploadResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MessageResponse":{"properties":{"message":{"type":"string","title":"Message"},"success":{"type":"boolean","title":"Success","default":true}},"type":"object","required":["message"],"title":"MessageResponse","description":"Simple message response"},"ResumeResponse":{"properties":{"id":{"type":"integer","title":"Id"},"filename":{"type":"string","title":"Filename"},"candidate_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Candidate Name"},"candidate_email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Candidate Email"},"candidate_phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Candidate Phone"},"skills":{"items":{"type":"string"},"type":"array","title":"Skills"},"experience_years":{"type":"number","title":"Experience Years"},"education":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Education"},"is_duplicate":{"type":"string","title":"Is Duplicate"},"uploaded_at":{"type":"string","format":"date-time","title":"Uploaded At"}},"type":"object","required":["id","filename","candidate_name","candidate_email","candidate_phone","skills","experience_years","education","is_duplicate","uploaded_at"],"title":"ResumeResponse"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"user":{"$ref":"#/components/schemas/UserResponse"}},"type":"object","required":["access_token","user"],"title":"TokenResponse","description":"Returned after successful login"},"UserLogin":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserLogin","description":"Data needed to login","example":{"email":"hr@company.com","password":"strongpassword123"}},"UserRegister":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"full_name":{"type":"string","title":"Full Name"},"password":{"type":"string","title":"Password"},"role":{"$ref":"#/components/schemas/UserRole"}},"type":"object","required":["email","full_name","password","role"],"title":"UserRegister","description":"Data needed to register a new user","example":{"email":"hr@company.com","full_name":"John Smith","password":"strongpassword123","role":"hr"}},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"email":{"type":"string","title":"Email"},"full_name":{"type":"string","title":"Full Name"},"role":{"$ref":"#/components/schemas/UserRole"},"is_active":{"type":"boolean","title":"Is Active"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","email","full_name","role","is_active","created_at"],"title":"UserResponse","description":"What we send back about a user — NEVER include password!"},"UserRole":{"type":"string","enum":["hr","student"],"title":"UserRole"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}